requirements.txt
python-pptx==0.6.18
pandas>=0.18.0
openpyxl>=2.4.7
six==1.12.0
